requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.testing.in, then run this to update:
#
#    pip-compile requirements.in requirements.testing.in -o requirements.txt
#

apipkg==1.4               # via execnet
click==6.2
cov-core==1.15.0          # via pytest-cov
coverage==4.0.3           # via cov-core
enum34==1.1.2
execnet==1.4.1            # via pytest-xdist
gnureadline==6.3.3        # via ipython
ipython==2.3.1
logilab-astng==0.24.3     # via pylint
logilab-common==1.1.0     # via logilab-astng, pylint
mock==1.0.1
numpy==1.10.4
pandas==0.16.2
py==1.4.31                # via pytest, pytest-xdist
pylint==0.25.1
pytest-cov==1.7.0
pytest-xdist==1.14
pytest==2.5.2
python-dateutil==2.4.1    # via pandas
python-igraph==0.7.1.post6
pytz==2014.10             # via pandas
scipy==0.15.1
six==1.10.0               # via logilab-common, python-dateutil, theano
theano==0.8.1

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via logilab-common
